% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/belief.R
\name{belief_update}
\alias{belief_update}
\title{Update beliefs from observations stored in a data frame}
\usage{
belief_update(d, cue, categories, beliefs)
}
\arguments{
\item{d}{Data frame with cue values and categories}

\item{cue}{Quoted name of column of data frame with cue values}

\item{categories}{Quoted name of column with known category labels.}

\item{beliefs}{Named list of starting belief NIX^2 parameters. Names must
match unique values of \code{d[[categories]]}}
}
\value{
The input data frame with parameter values updated after each
  observation in the \code{params} columns
}
\description{
Update beliefs from observations stored in a data frame
}
\seealso{
\code{\link{belief_update_batch}} for batch belief updating, which
  is more efficient.
}

