% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beliefupdatr-package.R
\docType{package}
\name{beliefupdatr-package}
\alias{beliefupdatr-package}
\alias{beliefupdatr}
\title{The 'beliefupdatr2' package.}
\description{
Modeling phonetic adaptation as belief updating requires some
  distributions and utility functions. Here they are.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
