% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stans.R
\name{compile_stan}
\alias{compile_stan}
\title{Compile an included Stan model}
\usage{
compile_stan(model_name)
}
\arguments{
\item{model_name}{(Relative) filename of model}
}
\description{
Compile an included Stan model
}

