% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nix.R
\name{d_nix2_predict}
\alias{d_nix2_predict}
\title{Predictive distribution for Normal-Chi^-2 distribution}
\usage{
d_nix2_predict(x, p, log = FALSE)
}
\arguments{
\item{x}{Point to evaluate likelihood at. Can be a vector.}

\item{p}{Parameters of Normal-Chi^-2 distribution.}

\item{log}{Whether to return log likelihood (default: FALSE)}
}
\description{
This is the probability distribution of observations drawn from a normal
distribution whose mean and variance follow the Normal-Chi^-2 distribution
with the specified parameters. It works out to be a Student's t distribution.
}
