% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles.R
\name{d_predict_particle}
\alias{d_predict_particle}
\alias{d_predict_particles}
\title{Posterior predictive function for a single particle}
\usage{
d_predict_particle(x, particle, log = FALSE)

d_predict_particles(x, particles, log = FALSE)
}
\arguments{
\item{x}{Points to calculate posterior predictive density at}

\item{particle}{The particle}

\item{log}{=FALSE}

\item{particles}{List of particles in population (weights must sum to 1)}
}
\value{
A named list of predictive (log-)densities for each category
}
\description{
Posterior predictive function for a single particle
}
\section{Functions}{
\itemize{
\item \code{d_predict_particles}: Posterior predictive distribution for
population of particles
}}

