% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nix.R
\name{d_scaled_inv_chisq}
\alias{d_scaled_inv_chisq}
\title{Scaled inverse chi-squared density}
\usage{
d_scaled_inv_chisq(y, sigma2, nu, log = FALSE)
}
\arguments{
\item{y}{Point to calculate density for}

\item{sigma2}{Scale parameter}

\item{nu}{Degrees of freedom}

\item{log}{Default: FALSE}
}
\value{
The density of y under the scaled-inverse-chi-squared distribution
}
\description{
We compute this based on the chi-squared density function.  Our random
variable y is obtained from a chi-squared(nu) variable x via
}
\details{
y = g(x) = nu * sigma^2 * 1/x
  x = g^-1(y) = nu*sigma^2 * 1/y

Thus, the density function f_Y(y) is obtained from f_X = dchisq via

f_Y(y) = d/dy g^-1(y) * f_X(g^-1(y))
         = y^-2 * nu*sigma^2 * dchisq(nu*sigma^2/y, nu)
}

