% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-conj-update.R
\name{extract_prior_samples}
\alias{extract_prior_samples}
\title{Extract sampled prior beliefs from fitted model to a data_frame}
\usage{
extract_prior_samples(fitted, cat_levels,
  samples = rstan::extract(fitted))
}
\arguments{
\item{fitted}{Fitted stan model for inferring prior}

\item{cat_levels}{Labels for the categories.}

\item{samples}{Optional: Samples extracted from fitted model with
\code{rstan::extract}, will override fitted if provided.}
}
\value{
a data_frame, with one row per sample and columns for
  \code{iterations} (sample number), \code{category}, \code{mu_0},
  \code{sigma_0}, \code{kappa_0}, and \code{nu_0}.
}
\description{
Extract sampled prior beliefs from fitted model to a data_frame
}
