% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-conj-update.R
\name{extract_updated_samples}
\alias{extract_updated_samples}
\title{Extract sampled updated beliefs from fitted model to data_frame}
\usage{
extract_updated_samples(fitted, cat_levels, group_levels,
  samples = rstan::extract(fitted))
}
\arguments{
\item{fitted}{Fitted stan model for inferring prior}

\item{cat_levels}{Labels for the categories.}

\item{group_levels}{Labels for groups/conditions (for updated)}

\item{samples}{Optional: Samples extracted from fitted model with
\code{rstan::extract}, will override fitted if provided.}
}
\value{
a data_frame, with one row per sample and columns for
  \code{iterations} (sample number), \code{category}, \code{group},
  \code{mu_n}, \code{sigma_n}, \code{kappa_n}, and \code{nu_n}.
}
\description{
Extract sampled updated beliefs from fitted model to data_frame
}
