% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-conj-update.R
\name{infer_prior_beliefs}
\alias{infer_prior_beliefs}
\title{Infer prior beliefs based on adaptation behavior}
\usage{
infer_prior_beliefs(df, cue, category, response, condition, ranefs,
  n_blocks, ...)
}
\arguments{
\item{df}{Data frame with adaptation data}

\item{cue}{(Quoted) name of columns in training and test which have the cue
values.}

\item{category}{(Quoted) name of column in training data with the correct
category labels.}

\item{response}{(Quoted) name of column in test data with responses}

\item{condition}{(Quoted) name of column with group identifiers}

\item{ranefs}{(Quoted) name of column(s) with random effect grouping
variables (like subject IDs).}

\item{n_blocks}{Number of blocks to divide data into.}

\item{...}{Additional parameters are passed to \code{\link[rstan]{sampling}}}
}
\value{
A \code{stanfit} object with the fitted stan model.
}
\description{
This takes training and test data, the names of the columns, and (optionally)
the number of blocks to split the trials into, and stan parameters, and uses
stan to draw samples from the prior beliefs that match the behavior.
}
\seealso{
\code{\link{extract_prior_samples}} to get a data frame of samples
  from the prior belief parameters.
}
