% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles.R
\name{init_particle}
\alias{init_particle}
\alias{init_particles}
\title{Initialize one particle based on NIX2 parameters}
\usage{
init_particle(params)

init_particles(params, n)
}
\arguments{
\item{params}{NIX2 parameters}

\item{n}{Population size}
}
\value{
A (population of) particle(s)
}
\description{
Initialize one particle based on NIX2 parameters
}
\section{Functions}{
\itemize{
\item \code{init_particles}: Initialize a population of particles
}}

