% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nix.R
\name{is_nix2_params}
\alias{is_nix2_params}
\title{Check parameter vector for Normal-Chi^2}
\usage{
is_nix2_params(p)
}
\arguments{
\item{p}{Named list to check}
}
\description{
Valid parameter vector has fields mu (expected mean), sigma2 (expected
variance), kappa (mean pseudocount), and nu (variance pseudocount/degrees of
freedom).
}
\details{
Kappa, nu, and sigma2 must be non-negative.
}

