% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-conj-update.R
\name{melt_samples}
\alias{melt_samples}
\title{Melt a mult-dimensional array of samples into a df}
\usage{
melt_samples(samples, varname)
}
\arguments{
\item{samples}{Named list of samples (like returned by rstan::extract)}

\item{varname}{Name of variable whose samples to melt}
}
\value{
A tidy dataframe of samples
}
\description{
Melt a mult-dimensional array of samples into a df
}
