% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stans.R
\name{model_filename}
\alias{model_filename}
\title{Get a model's full path}
\usage{
model_filename(model_name)
}
\arguments{
\item{model_name}{(Relative) filename of model}
}
\description{
Code for stan models is stored in the exec/ subdirectory of wherever
this package is installed. This is a convenience shortcut to get fully
specified paths to a model
}
