% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nix.R
\name{nix2_classify}
\alias{nix2_classify}
\title{Classification function for Normal-Chi^-2 distributions}
\usage{
nix2_classify(x, ps, category = 1:length(ps))
}
\arguments{
\item{x}{Vector of points to classify}

\item{ps}{List of category Norma-Chi^-2 params}

\item{category}{(Optional) Which categories' posteriors to return. If
missing, all are returned}
}
\value{
A length(x)-by-category matrix of posterior probabilities
}
\description{
This computes the posterior probability of data under different Normal-Chi^-2
distributions.
}

