% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nix.R
\name{nix2_log_norm_const}
\alias{nix2_log_norm_const}
\title{Normalizing constant for NIX^2 distribution}
\usage{
nix2_log_norm_const(p)
}
\arguments{
\item{p}{NIX^2 parameters}
}
\description{
The terms of \eqn{p(\mu, \sigma^2 | \mu_0, \sigma^2_0, \nu, \kappa)} that
don't involve \eqn{\mu} or \eqn{\sigma^2}.  Useful in calculating the
marginal likelihood of data, where the mean and variance are integrated out.
}
