% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nix.R
\name{nix2_marginal_lhood}
\alias{nix2_marginal_lhood}
\title{Marginal likelihood function}
\usage{
nix2_marginal_lhood(p, x = NA, n = length(x), xbar = mean(x), ss = s2 *
  (n - 1), s2 = var(x), log = FALSE)
}
\arguments{
\item{p}{named list of distribution parameters (mu, sigma2, nu, and kappa)}

\item{x}{vector of observations to update with (optional if summary stats
are passed)}

\item{n}{number of observations (overrides x)}

\item{xbar}{observed mean (overrides x)}

\item{ss}{observed sum of squares (overrides x and s2)}

\item{s2}{observed sample variance (ss / (n-1); overrides x)}

\item{log}{Default FALSE.}
}
\value{
The overal marginal likelihood of all the observations x (or summary
  statistics if provided).
}
\description{
Compute the marginal likelihood of data given NIX^2 prior, or \eqn{p(x) =
\int \mathrm d \sigma_2 \mathrm d\mu p(x | \mu, \sigma^2) p(\mu, \sigma^2)}.
}
\details{
Based on equation 171 from
\href{https://www.cs.ubc.ca/~murphyk/Papers/bayesGauss.pdf}{Murphy (2008;
University of British Columbia)}.
}
