% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nix.R
\name{nix2_params}
\alias{nix2_params}
\title{Wrap parameters into a list for updating etc.}
\usage{
nix2_params(...)
}
\arguments{
\item{mu}{Mean}

\item{sigma2}{Variance}

\item{kappa}{Mean prior pseudocount}

\item{nu}{Variance prior pseudocount}
}
\description{
Wrap parameters into a list for updating etc.
}

