% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nix.R
\name{nix2_update}
\alias{nix2_update}
\title{Conjugate updating of Normal-Chi^-2 parameters}
\usage{
nix2_update(p = list(nu = 0, kappa = 0, mu = 0, sigma2 = 0), x = NA,
  n = length(x), xbar = mean(x), ss = s2 * (n - 1), s2 = var(x))
}
\arguments{
\item{p}{named list of distribution parameters (mu, sigma2, nu, and kappa)}

\item{x}{vector of observations to update with (optional if summary stats
are passed)}

\item{n}{number of observations (overrides x)}

\item{xbar}{observed mean (overrides x)}

\item{ss}{observed sum of squares (overrides x and s2)}

\item{s2}{observed sample variance (ss / (n-1); overrides x)}
}
\value{
Updated parameters
}
\description{
This uses the parametrization from
\href{https://www.cs.ubc.ca/~murphyk/Papers/bayesGauss.pdf}{Murphy (2008;
University of British Columbia)}.
}

