% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nix.R
\name{nix2_update_one}
\alias{nix2_update_one}
\title{Conjugate updating of Normal-Chi^-2 parameters for a single observation}
\usage{
nix2_update_one(p = list(nu = 0, kappa = 0, mu = 0, sigma2 = 0), x)
}
\arguments{
\item{p}{named list of distribution parameters (mu, sigma2, nu, and kappa)}

\item{x}{vector of observations to update with (optional if summary stats
are passed)}
}
\value{
Update parameters
}
\description{
This uses the parametrization from
\href{https://www.cs.ubc.ca/~murphyk/Papers/bayesGauss.pdf}{Murphy (2008;
University of British Columbia)}.
}
\details{
This is specialized for a single observation, and about 25% faster than
\code{nix2_update}. Use, for instance, in \code{reduce} or \code{accumulate}.
}
