% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles.R
\name{particle_params_to_nix2}
\alias{particle_params_to_nix2}
\title{Convert prior + sufficient stats to NIX2 params}
\usage{
particle_params_to_nix2(params)
}
\description{
Convert prior + sufficient stats to NIX2 params
}

