% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-conj-update.R
\name{prepare_data_conj_infer_prior}
\alias{prepare_data_conj_infer_prior}
\title{Convert data to stan input for conjugate prior inference}
\usage{
prepare_data_conj_infer_prior(training, test, cue, category, response, subject)
}
\arguments{
\item{training}{Data frame with training data: 'cue', 'category', and
'subject'.}

\item{test}{Data frame with test data: 'cue', 'response', and 'subject'}

\item{cue}{(Quoted) name of columns in training and test which have the cue
values.}

\item{category}{(Quoted) name of column in training data with the correct
category labels.}

\item{response}{(Quoted) name of column in test data with responses}

\item{subject}{(Quoted) name of column with subject identifiers}
}
\value{
a list with elements: x, y, z (training cues, subjects, and
  categories), n, m, l (number of observations, subject, and categories),
  x_test, y_test, z_test, and n_test.
}
\description{
See exec/conj_id_lapsing_fit.stan for data format expected
}
\examples{
\dontrun{
supunsup::supunsup_clean \%>\%
  filter(supCond == 'unsupervised') \%>\%
  prepare_data_conj_infer_prior(training=., test=.,
                                'vot', 'trueCat', 'respCat', 'subject')
}
}

