% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-conj-update.R
\name{prepare_data_conj_suff_stats_infer_prior}
\alias{prepare_data_conj_suff_stats_infer_prior}
\title{Convert data into format for bulk updating with sufficient statistics}
\usage{
prepare_data_conj_suff_stats_infer_prior(training, test, cue, category,
  response, group)
}
\arguments{
\item{training}{Data frame with training data: 'cue', 'category', and
'subject'.}

\item{test}{Data frame with test data: 'cue', 'response', and 'subject'}

\item{cue}{(Quoted) name of columns in training and test which have the cue
values.}

\item{category}{(Quoted) name of column in training data with the correct
category labels.}

\item{response}{(Quoted) name of column in test data with responses}

\item{group}{(Quoted) name of column with group identifiers}
}
\value{
A list of data for 'conj_id_lapsing_sufficient_stats_fit.stan'.
}
\description{
Like \code{\link{prepare_data_conj_infer_prior}}, but for sufficient
statistics for training data (count, mean, and sample standard deviation)
instead of raw training observations.
}

