% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nix.R
\name{r_nix2}
\alias{r_nix2}
\title{Sample from NIX2 distribution}
\usage{
r_nix2(n, p)
}
\arguments{
\item{n}{Number of samples to draw}

\item{p}{NIX2 parameters}
}
\value{
An n x 2 matrix with sampled means in the first column and variances
  in the second
}
\description{
This first draws the variance from a (scaled) inverse chi squared
distribution, and then the mean from a normal distribution.
}
