% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles.R
\name{d_predict_particle}
\alias{d_predict_particle}
\alias{d_predict_particles_df}
\alias{d_predict_particles_marginal}
\alias{d_predict_particles_marginal_df}
\title{Posterior predictive density functions}
\usage{
d_predict_particle(x, particle, log = FALSE)

d_predict_particles_marginal(x, particles, log = FALSE)

d_predict_particles_df(x, particles, log = FALSE)

d_predict_particles_marginal_df(x, particles, log = FALSE)
}
\arguments{
\item{x}{Points to calculate posterior predictive density at}

\item{particle}{The particle}

\item{log}{=FALSE}

\item{particles}{List of particles in population (weights must sum to 1)}
}
\value{
A named list of predictive (log-)densities for each category, or a
  data_frame with columns x, category, pred (density), particle_id, and
  weight
}
\description{
Posterior predictive distribution for a single particle
}
\section{Functions}{
\itemize{
\item \code{d_predict_particles_marginal}: Marginal posterior predictive distribution for
population of particles (the weighted average of the individual particles'
posterior predictive distributions)

\item \code{d_predict_particles_df}: Posterior predictive densities for all
particles in a population, organized into a data_frame.

\item \code{d_predict_particles_marginal_df}: Marginal posterior predictive distribution
function for a whole population, as a data_frame
}}

