% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-conj-update.R
\name{prepare_data_conj_suff_stats_infer_prior}
\alias{prepare_data_conj_suff_stats_infer_prior}
\title{Convert data into format for bulk updating with sufficient statistics}
\usage{
prepare_data_conj_suff_stats_infer_prior(df, cue, category, response,
  condition, ranefs, test_df = df)
}
\arguments{
\item{df}{Data frame with adaptation data}

\item{cue}{(Quoted) name of columns in training and test which have the cue
values.}

\item{category}{(Quoted) name of column in training data with the correct
category labels.}

\item{response}{(Quoted) name of column in test data with responses}

\item{condition}{(Quoted) name of column with group identifiers}

\item{ranefs}{(Quoted) name of column(s) with random effect grouping
variables (like subject IDs).}

\item{test_df}{(optional) data frame with test data (all data is used as test
by default).}
}
\value{
A list of data for 'conj_id_lapsing_sufficient_stats_fit.stan'.
}
\description{
For Stan models that are based on sufficient statistics for training data
(count, mean, and sample standard deviation) instead of raw training
observations.
}
